Use SQLAlchemy to handle database connections.
With SQLAlchemy, pandasql can use a variety of backends to offload
computation to.
A use case for this would be working with data that does not fit in
memory, or queries that produce (intermediate) result sets that do not
fit in memory.

Backward compatibility to the old dbapi backend is maintained, both
for in memory and on disk sqlite persistence.